requirements.txt
streamlit==1.38.0
PyPDF2==3.0.1
google.generativeai==0.8.2
python-dotenv==1.0.1